pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "bs4>=0.0.2",
    "lxml>=6.0.2",
    "pika>=1.3.2",
    "prometheus-client>=0.23.1",
    "psycopg2>=2.9.11",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.6",
    "sqlalchemy>=2.0.44",
]
